Thread pools
we have SomeThread class which extends Thread. Inside the App class, we are creating a fixed thread pool from Executors framework. Then we pass the object of thread type for executing threads.  And shut down the Executor pool.  shutdown() method helps us to finish the execution of all the submitted tasks and terminate the thread pool.