Blocking_Queue module
he blocking queue is used as a container for objects when the threads are inserting the objects and other threads are taking the objects out of them. The threads can only continue inserting the objects until a specific limit and take out or use the resources until a specific limit. Therefore, any limit reached by the blocking queue will immediately block the threads.

The following program will demonstrate the capability of blocking queues. In order to understand the concept of blocking queue, we will start with the producer and consumer. As the name suggests producer thread adds the element in the queue and the consumer thread consumes the element from the queue. These two thread works simultaneously on blocking queue collection.

We have created a producer class that implements the Runnable interface. The producer class will use the put() method to add elements in the queue and before putting a new element the thread will wait for one second.

Once the element in the blocking queue reached the maximum limit, instead of terminating the program blocking queue will force the thread to wait until there is an empty space.